Extracted UI-specific startup helpers and classes into UI project
