+++ src/Skoruba.IdentityServer4.Admin.UI/Configuration/AdminConfiguration.cs
