+++ src/Skoruba.IdentityServer4.Admin.UI/Configuration/ApplicationParts/GenericControllerRouteConvention.cs
