+++ src/Skoruba.IdentityServer4.Admin.UI/Configuration/ApplicationParts/GenericTypeControllerFeatureProvider.cs
