+++ src/Skoruba.IdentityServer4.Admin.UI/Configuration/Interfaces/IRootConfiguration.cs
